reorganize: move classes that deal with templates
