+++ src/dct/Theater.lua
@@ -15,9 +15,9 @@ local dctutils    = require("dct.utils")
 local uicmds      = require("dct.ui.cmds")
 local uiscratchpad= require("dct.ui.scratchpad")
 local Observable  = require("dct.Observable")
-local STM         = require("dct.STM")
-local Template    = require("dct.Template")
-local Region      = require("dct.Region")
+local STM         = require("dct.templates.STM")
+local Template    = require("dct.templates.Template")
+local Region      = require("dct.templates.Region")
 local Asset       = require("dct.Asset")
 local AssetManager= require("dct.AssetManager")
 local Commander   = require("dct.ai.Commander")
